CB#16 - deposit address - Rename folder Jobs -> InitializationTasks
